+++ Vault-Coder/src/utopia/vault/coder/model/scala/template/CodeConvertible.scala
@@ -1,17 +1,10 @@
 package utopia.vault.coder.model.scala.template
 
 import utopia.flow.util.FileExtensions._
+import utopia.vault.coder.model.scala.Code
 
 import java.nio.file.Path
 import scala.io.Codec
-
-object CodeConvertible
-{
-	/**
-	  * Recommended maximum line length
-	  */
-	val maxLineLength = 100
-}
 
 /**
   * Common trait for items which can be converted to 1 or more full lines of code
@@ -23,10 +16,10 @@ trait CodeConvertible
 	// ABSTRACT ------------------------------
 	
 	/**
-	  * @return Code lines based on this item. Expects the topmost line not to be intended but the other
+	  * @return This item converted to code. Expects the topmost line not to be intended but the other
 	  *         lines to be intended correctly relative to the topmost line.
 	  */
-	def toCodeLines: Vector[String]
+	def toCode: Code
 	
 	
 	// OTHER    -----------------------------
@@ -38,5 +31,5 @@ trait CodeConvertible
 	  * @return Success or failure
 	  */
 	def writeTo(path: Path)(implicit codec: Codec) =
-		path.createParentDirectories().flatMap { _.writeLines(toCodeLines) }
+		path.createParentDirectories().flatMap { _.writeLines(toCode.lines.map { _.toString }) }
 }